[ll][gl] moved the queue implementation into a module
